# pqarrow: more optimizations on top of ParquetConverter