The first two commits are #195. The other commits have been split out into this PR to make reviewing that easier, so if you haven't already, please review #195. The additional performance benefits this PR offers us are:
```
name          old time/op    new time/op    delta
QueryTypes-8    5.31ms ± 1%    4.87ms ± 1%  -8.26%  (p=0.008 n=5+5)
QueryMerge-8    18.7ms ± 0%    18.5ms ± 3%     ~     (p=0.151 n=5+5)
QueryRange-8    11.3ms ± 2%    10.6ms ± 3%   -6.02%  (p=0.016 n=4+5)

name          old alloc/op   new alloc/op   delta
QueryTypes-8    5.62MB ± 0%    5.61MB ± 0%   -0.17%  (p=0.008 n=5+5)
QueryMerge-8    49.7MB ± 0%    43.0MB ± 0%  -13.51%  (p=0.008 n=5+5)
QueryRange-8    22.6MB ± 0%    15.9MB ± 0%  -29.65%  (p=0.008 n=5+5)

name          old allocs/op  new allocs/op  delta
QueryTypes-8     93.1k ± 0%     85.9k ± 0%   -7.64%  (p=0.008 n=5+5)
QueryMerge-8      146k ± 0%      143k ± 0%   -2.40%  (p=0.008 n=5+5)
QueryRange-8     49.2k ± 0%     45.7k ± 0%   -7.11%  (p=0.016 n=4+5)
```